+++ include/ignition/transport/NodeShared.hh
@@ -33,8 +33,8 @@
 #include <vector>
 #include <map>
 
+#include "ignition/transport/Export.hh"
 #include "ignition/transport/HandlerStorage.hh"
-#include "ignition/transport/Helpers.hh"
 #include "ignition/transport/Publisher.hh"
 #include "ignition/transport/RepHandler.hh"
 #include "ignition/transport/ReqHandler.hh"
@@ -266,11 +266,6 @@ namespace ignition
 
       /// \brief When true, the reception thread will finish.
       public: bool exit;
-
-#ifdef _WIN32
-      /// \brief True when the reception thread is finishing.
-      public: bool threadReceptionExiting;
-#endif
 
       /// \brief Port used by the message discovery layer.
       private: const int kMsgDiscPort = 11317;