+++ include/ignition/transport/SubscriptionHandler.hh
@@ -40,7 +40,7 @@
 
 #include <ignition/msgs/Factory.hh>
 
-#include "ignition/transport/Helpers.hh"
+#include "ignition/transport/Export.hh"
 #include "ignition/transport/MessageInfo.hh"
 #include "ignition/transport/SubscribeOptions.hh"
 #include "ignition/transport/TransportTypes.hh"