+++ test/integration/twoProcessesSrvCallReplierIncreasing_aux.cc
@@ -31,11 +31,10 @@ static int Forever = INT_MAX;
 
 //////////////////////////////////////////////////
 /// \brief Provide a service.
-void srvEcho(const ignition::msgs::Int32 &_req, ignition::msgs::Int32 &_rep,
-  bool &_result)
+bool srvEcho(const ignition::msgs::Int32 &_req, ignition::msgs::Int32 &_rep)
 {
   _rep.set_data(_req.data());
-  _result = true;
+  return true;
 }
 
 //////////////////////////////////////////////////