+++ Changelog.md
@@ -3,6 +3,11 @@
 1. Added publishing and receiving messages as raw bytes
     * [Pull request 251](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/251)
 
+1. Updated service responder callbacks to return a boolean value. The
+   existing functions have been deprecated.
+    * [Pull request 260](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/260)
+    * [Pull request 228](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/228)
+
 ## Ignition Transport 3.0.0
 
 1. Added optional message throttling when publishing messages.